requirements.txt
###### Requirements without Version Specifiers ######
prometheus_client
pyserial
sdnotify
